feat(infra): implement Postgres persistence layer for `DetailTransaction`
- Added `DetailTransactionRepository` and `DetailTransactionAdapter` for data access.
- Introduced `IDetailTransactionEntityMapper` for model-entity mapping.
- Updated `DetailTransactionEntity` to use `Double` for `unitPrice`.
- Adjusted fields in `DetailTransaction` domain model.

+++ src/main/java/com/lian/marketing/transactionmicroservice/infrastructure/driven/r2dbc/postgres/entity/DetailTransactionEntity.java
@@ -18,7 +18,7 @@ public class DetailTransactionEntity {
     @Id
     private UUID id;
     @Column("unit_price")
-    private Integer unitPrice;
+    private Double unitPrice;
     @Column("quantity")
     private Integer quantity;
     @Column("transaction_id")
